Remove birdkeeper
Make `pairrit` responsible for looking up and modifying pairs, and
remove birdkeeper as it is no longer necessary.
