+++ lib/index.js
@@ -1,6 +1,6 @@
 const express = require('express');
 const app = express();
-const keeper = require('./birdkeeper');
+const pairrit = require('./pairrit');
 const pluralize = require('pluralize');
 
 let state = {}
@@ -41,20 +41,20 @@ const commands = {
   },
 
   join: function(req, res) {
-    state = keeper.update(state, req.options);
+    state = pairrit.update(state, req.options);
 
     res.send(`You have joined the \`${req.options.pairName}\` pair.`);
   },
 
   leave: function(req, res) {
-    state = keeper.update(state, req.options);
+    state = pairrit.update(state, req.options);
 
     res.send(`You have left the \`${req.options.pairName}\` pair.`);
   },
 
   list: function(req, res) {
     {
-      const pairs = keeper.list(state, req.options.pairChannel);
+      const pairs = pairrit.list(state, req.options.pairChannel);
 
       if (pairs.length > 0) {
         const message = { text: 'The following pairs were found:', attachments: [] };
@@ -80,7 +80,7 @@ const commands = {
   default: function(req, res) {
     res.send(`The provided command (\`/pair ${command}\`) was not recognized.\nPerhaps try \`/pair help\` for a list of valid commands.`);
   }
-}
+};
 
 const router = function router(req, res) {
   const args = req.query.text.split(' ');
